@@ -1,8 +1,14 @@
 Spree::User.class_eval do
 
-  has_many :affiliates, :class_name => 'Spree::Affiliate', :foreign_key => "partner_id"
-  has_many :referred_users, :class_name => 'Spree::User', :through => :affiliates
-  has_one :affiliate_partner, :class_name => "Spree::Affiliate", :foreign_key => "user_id"
+  has_many :affiliates,
+    class_name: 'Spree::Affiliate',
+    foreign_key: 'partner_id'
+  has_many :referred_users,
+    class_name: 'Spree::User',
+    through: :affiliates
+  has_one :affiliate_partner,
+    class_name:'Spree::Affiliate',
+    foreign_key: 'user_id'
 
   def referred_by
     affiliate_partner.partner
